pyproject.toml
[project]
name = "ai-quiz-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "chromadb>=1.1.1",
    "fastapi>=0.118.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]
